backend/app/serializers/category_serializer.py
from rest_framework import serializers

from app.models import Category


# CategorySerializer
class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'category_name']
