backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-extra-fields==3.0.2
django-filter==23.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==19.6.2
Markdown==3.4.3
psycopg2-binary==2.9.7
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
